refactor: Now using ContentHandler internally for parsing
Before the ContentHandler was provided as an extra feature
but was not actualy used by the code itself. This meant there
actually existed two very similar parsers. The parser not
using the ContentHandler was removed. And the ContentHandler
API was improved to make it easier to write simple handlers
that don't need to do a lot of state mangement by passing
the current parse state to them.